+++ flutter_google_places_sdk_platform_interface/lib/flutter_google_places_sdk_platform.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_google_places_sdk_platform_interface/method_chanel_flutter_google_places_sdk.dart';
+import 'package:flutter_google_places_sdk_platform_interface/flutter_google_places_sdk_platform_interface.dart';
+import 'package:flutter_google_places_sdk_platform_interface/method_channel_flutter_google_places_sdk.dart';
+import 'package:flutter_google_places_sdk_platform_interface/src/types/fetch_place_response.dart';
 import 'package:flutter_google_places_sdk_platform_interface/src/types/find_autocomplete_predictions_response.dart';
+import 'package:flutter_google_places_sdk_platform_interface/src/types/place_field.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 abstract class FlutterGooglePlacesSdkPlatform extends PlatformInterface {
@@ -30,18 +33,29 @@ abstract class FlutterGooglePlacesSdkPlatform extends PlatformInterface {
     throw UnimplementedError('deinitialize() has not been implemented.');
   }
 
-  Future<void> initialize(String apiKey, {Locale locale}) {
+  Future<void> initialize(String apiKey, {Locale? locale}) {
     throw UnimplementedError('initialize() has not been implemented.');
   }
 
-  Future<bool> isInitialized() {
+  Future<bool?> isInitialized() {
     throw UnimplementedError('isInitialized() has not been implemented.');
   }
 
   Future<FindAutocompletePredictionsResponse> findAutocompletePredictions(
-      String query,
-      {List<String> countries}) {
+    String query, {
+    List<String>? countries,
+    bool? newSessionToken,
+    LatLng? origin,
+  }) {
     throw UnimplementedError(
         'findAutocompletePredictions() has not been implemented.');
+  }
+
+  Future<FetchPlaceResponse> fetchPlace(
+    String placeId, {
+    List<PlaceField>? fields,
+    bool? newSessionToken,
+  }) {
+    throw UnimplementedError('fetchPlaceDetails() has not been implemented.');
   }
 }