+++ flutter_google_places_sdk_platform_interface/lib/src/types/find_autocomplete_predictions_response.dart
@@ -18,7 +18,6 @@ class FindAutocompletePredictionsResponse {
   int get hashCode => predictions.hashCode;
 
   @override
-  String toString() {
-    return 'FindAutocompletePredictionsResponse{predictions: $predictions}';
-  }
+  String toString() =>
+      'FindAutocompletePredictionsResponse{predictions: $predictions}';
 }