+++ flutter_google_places_sdk_platform_interface/lib/src/types/utils.dart
@@ -1,0 +1,7 @@
+Map<String, dynamic>? toJsonMap(dynamic value) {
+  if (value == null) {
+    return null;
+  }
+
+  return Map<String, dynamic>.from(value);
+}