+++ flutter_google_places_sdk_platform_interface/test/method_channel_flutter_google_places_sdk_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_google_places_sdk_platform_interface/flutter_google_places_sdk_platform_interface.dart';
-import 'package:flutter_google_places_sdk_platform_interface/method_chanel_flutter_google_places_sdk.dart';
+import 'package:flutter_google_places_sdk_platform_interface/method_channel_flutter_google_places_sdk.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
@@ -58,9 +58,13 @@ void main() {
     test('findAutocompletePredictions', () async {
       const testQuery = 'my-test-query';
       const testCountries = ['c1', 'c2'];
+      const newSessionToken = true;
+      const origin = LatLng(lat: 325.21, lng: -952.52);
       await places.findAutocompletePredictions(
         testQuery,
         countries: testCountries,
+        newSessionToken: newSessionToken,
+        origin: origin,
       );
       expect(
         log,
@@ -69,7 +73,30 @@ void main() {
               arguments: <String, Object>{
                 'query': testQuery,
                 'countries': testCountries,
+                'newSessionToken': newSessionToken,
+                'origin': origin.toMap()
               })
+        ],
+      );
+    });
+
+    test('fetchPlace', () async {
+      const placeId = 'my-test-place-id';
+      const testFields = [PlaceField.Location, PlaceField.Address];
+      const newSessionToken = true;
+      await places.fetchPlace(
+        placeId,
+        fields: testFields,
+        newSessionToken: newSessionToken,
+      );
+      expect(
+        log,
+        <Matcher>[
+          isMethodCall('fetchPlace', arguments: <String, Object>{
+            'placeId': placeId,
+            'fields': testFields.map((e) => e.value).toList(growable: false),
+            'newSessionToken': newSessionToken,
+          })
         ],
       );
     });